requirements.txt
django==5.2.6
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.3.0
django-widget-tweaks==1.5.0
dj-database-url==2.1.0
